+++ CHANGELOG.md
@@ -19,6 +19,8 @@ This project follows [Keep a Changelog](https://keepachangelog.com/) and [Semant
 ### Added
 - HTML comments to ignore secret detection in HTML lines (e.g. `<!-- dotenv-diff-ignore -->`).
 - Also ignore html sections with `<!-- dotenv-diff-ignore-start -->` and `<!-- dotenv-diff-ignore-end -->`.
+- Added option to have a dotenv-diff.config.json file for configuration.
+- ignoreUrls option to ignore specific URLs in secret detection. (e.g. `https://nomistake.com`).
 
 ## [2.2.8] - 2025-09-30
 ### Added